Separate out regenerate bundle image logic from build.py
Refers to CLOUDDST-5856
